+++ src/django_resume/plugins/__init__.py
@@ -1,6 +1,7 @@
 from .registry import plugin_registry
 from .about import AboutPlugin
 from .base import SimplePlugin, ListPlugin
+from .cover import CoverPlugin
 from .education import EducationPlugin
 from .identity import IdentityPlugin
 from .projects import ProjectsPlugin
@@ -11,6 +12,7 @@ from .timelines import EmployedTimelinePlugin, FreelanceTimelinePlugin
 
 __all__ = [
     "AboutPlugin",
+    "CoverPlugin",
     "EducationPlugin",
     "EmployedTimelinePlugin",
     "FreelanceTimelinePlugin",