+++ src/django_resume/views.py
@@ -45,5 +45,34 @@ def cv(request: HttpRequest, slug: str) -> HttpResponse:
     return render(request, "django_resume/plain/cv.html", context=context)
 
 
+def detail(request: HttpRequest, slug: str) -> HttpResponse:
+    resume = get_object_or_404(Resume.objects.select_related("owner"), slug=slug)
+
+    edit = bool(dict(request.GET).get("edit", False))
+    is_editable = request.user.is_authenticated and resume.owner == request.user
+    show_edit_button = True if is_editable and edit else False
+
+    edit_url, show_url = get_edit_and_show_urls(request)
+    context = {
+        "resume": resume,
+        # needed to include edit styles in the base template
+        "show_edit_button": show_edit_button,
+        "is_editable": is_editable,
+        "edit_url": edit_url,
+        "show_url": show_url,
+    }
+    plugin_names = ["about", "identity", "cover"]
+    for name in plugin_names:
+        plugin = plugin_registry.get_plugin(name)
+        context[plugin.name] = plugin.get_context(
+            request,
+            plugin.get_data(resume),
+            resume.pk,
+            context={},
+            edit=show_edit_button,
+        )
+    return render(request, "django_resume/plain/detail.html", context=context)
+
+
 def index(request):
     return render(request, "django_resume/plain/index.html")